Added "ACCEPTED" and "DROPPED" query conditions representing the set of all accepted packets and the set of all dropped packets, respectively (imagine that!)
Also completely changed the syntax of Assertions and Queries.  The input chain is now
a parameter to the "LOGGED", "ACCEPTED", and "DROPPED" conditions.  This means that all example
query files will need to be modified.

History is now working, I believe, but there is no way to access it, at the moment, except by hacking
the code to temporarily call FW->HistoryForest->PrintMDD().